detect and report error
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@59173 c90b9560-bf6c-de11-be94-00142212c4b1

+++ pear/PEAR/Common.php
@@ -386,7 +386,9 @@ class PEAR_Common extends PEAR
         $xml = $pkgdir . DIRECTORY_SEPARATOR . 'package.xml';
 
         $tar = new Archive_Tar($file, true);
-        $tar->extractList($xml);
+        if (!$tar->extractList($xml)) {
+            return $this->raiseError('could not extract the package.xml file');
+        }
         $info = $this->infoFromDescriptionFile($xml);
         unlink($xml);
         return $info;
